vitpose_requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
backcall==0.2.0
blinker==1.8.2
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==42.0.6
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
dnspython==2.6.1
einops==0.3.0
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
gradio==3.13.2
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
invisible-watermark==0.2.0
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
jmespath==0.10.0
json-tricks==3.17.3
kiwisolver==1.4.5
kornia==0.6.0
lightning-utilities==0.11.2
linkify-it-py==2.0.3
Markdown==3.6
markdown-it-py==3.0.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mmcv-full==1.5.0
# Editable install with no version control (mmdet==2.26.0)
-e /dat03/xuanwenjie/code/tmp/mmdetection-2.26.0
mmengine==0.10.4
# Editable install with no version control (mmpose==0.24.0)
-e /dat03/xuanwenjie/code/animal_pose/ViTPose-main
# Editable install with no version control (mmsegmentation==0.30.0)
-e /dat03/maoyuanye/projects/FudanOCR-main/text-focused-Transformers
model-index==0.1.11
multidict==6.0.5
munkres==1.1.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
open-clip-torch==2.7.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.82
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.0
ordered-set==4.1.0
orjson==3.10.3
oss2==2.17.0
packaging==24.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==1.4.2
pytz==2023.4
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9
scipy==1.10.1
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
# Editable install with no version control (stable-diffusion==0.0.1)
-e /dat03/xuanwenjie/code/github_projects/HumanSD-main
stack-data==0.6.3
starlette==0.37.2
streamlit==1.35.0
streamlit-drawable-canvas==0.8.0
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
test-tube==0.7.5
timm==1.0.9
tomli==2.0.1
torch @ file:///dat03/maoyuanye/wheel/torch-1.10.0%2Bcu111-cp38-cp38-linux_x86_64.whl#sha256=a5a0fa82173439a693020b418370c7b683051410dc5eedb97331ca87af478f77
torchmetrics==0.6.0
torchvision @ file:///dat03/maoyuanye/wheel/torchvision-0.11.0%2Bcu111-cp38-cp38-linux_x86_64.whl#sha256=a6b0fd9f31513369f6c4cf5c7d2e54bebd8eafbdd169c68923207d7d7bfed5fd
tqdm==4.65.2
traitlets==5.14.3
triton==2.3.0
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==1.26.18
urwid==2.6.12
uvicorn==0.30.1
watchdog==4.0.1
wcwidth==0.2.13
webdataset==0.2.5
websockets==12.0
xtcocotools==1.14.3
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
